+++ lib/openhab/rspec/hooks.rb
@@ -80,6 +80,9 @@ module OpenHAB
           @metadata_provider.restore_parent
           @profile_factory_registration.unregister
           DSL::TimerManager.instance.cancel_all
+          # timers and rules have already been canceled, so we can safely just
+          # wipe this
+          DSL::Items::TimedCommand.timed_commands.clear
           Timecop.return
           restore_autoupdate_items
           Mocks::PersistenceService.instance.reset