+++ CHANGELOG.md
@@ -162,6 +162,10 @@ here is a non-exhaustive list of significant departures from the original gem:
   {OpenHAB::DSL.unit unit}, {OpenHAB::DSL.persistence persistence}, and
   {OpenHAB::DSL.ensure_states ensure_states}.
 * Fix thread safety issue that could cause timers to not be canceled when the script unloads.
+* {OpenHAB::DSL::Items::TimedCommand#command Timed command} thread safety issues resolved
+* {OpenHAB::DSL::Items::TimedCommand#command Timed command} now resets the on_expire setting when called again
+* {OpenHAB::DSL::Items::TimedCommand#command Timed command} still sends the command even the previous timed command is still pending
+* {OpenHAB::DSL::Items::TimedCommand#command Timed command} works with resetting to {NULL}/{UNDEF}
 
 ## [4.45.2](https://github.com/boc-tothefuture/openhab-jruby/compare/4.45.1...4.45.2) (2022-10-02)